Move status back to statusUtil
